cortexm: Improved connect-under-reset handling so we halt the core on the reset vector
When attaching under reset, it's critical that user code not be
allowed to run before the debugger commands the chip to halt.
If bad user code is loaded on the device (e.g. that disables the
clock), any window in between releasing reset and commanding a
halt can prevent the BMP from attaching. The BMP incorrectly
releases reset before commanding a halt (in cortexm_forced_halt),
meaning that a bad ELF file can prevent the BMP from attaching,
effectively bricking the chip.