use SWIM pins as USART
some STM ST-LINK/V2 also provide SWIM functionality to flash STM8.
the SWIM pins are connected to UART (USART1 REMAP):
- SWIM_RST is connected to PB6/USART1_TX
- SWIM is connected to PB7/USART1_RX (pulled up by 680R)

cheap ST-LINK/V2 clones in form of USB dongle with aluminium case
provide SWD and SWIM signals.
since blackmagic does not support SWIM functionality, the pins can
be re-used as UART port for the USB to UART bridge.

STM ST-LINK/V2 can also offer UART, using USART2, but then the
SWIM signals are not present.
this changes verifies if the SWIM signals are present and remaps
them to USART1 if this is the case.