Special case value-line F750 flash mapping
The value F750 part has the same idcode as a F745, but only has 64KiB of
flash. If we use the same flash mapping calculation for both, the
`remains` value will **underflow**, resulting in a flash size of
`0xfffd'0000` for the region starting at `0x8040'0000`. This causes GDB
to warn

    Overlapping regions in memory map: ignoring

and causes programming of the F750 part to fail.

This patch adds a check on the flash size of the device, and in the case
of the F750 adds the two smaller sectors.
Memory difference information can be seen on page 77 of ST manual RM0385
(Rev 8).