Probe LPC546xx before LPC43xx
When targeting LPC546xx, the adapter fails when probing for LPC43xx and
exits probing without trying LPC546xx. Not sure why that's happening. I
don't have an LPC43xx board to verify this change is OK there. I
verified the LPC546xx behavior is OK after this patch on an
LPCXpresso54628 board.