clang-format: add initial format file
Add a format file for `clang-format`. This is based on the
`.clang-format` file from Linux, adjusted to work with this project.

Notably, it opts for 120 characters, given that some functions are
rather wide.

Care was chosen to adopt settings that resulted in a minimum of
disruption when run through the formatter, however since no formatter
has been used before this necessarily results in code that does not
conform to the formatter.

Signed-off-by: Sean Cross <sean@xobs.io>