rtt_if: include "platform.h"
Include the `platform.h` file in order to read in various setteings such
as `RTT_UP_BUF_SIZE`, `RTT_DOWN_BUF_SIZE`, and `PC_HOSTED`.

These flags should all be set on the command line, however when using
this file under ccache these flags may not be set. As a result, ccache
may attempt to include `usbd.h` even when it is not preset, such as when
building in another repository.

Signed-off-by: Sean Cross <sean@xobs.io>