gdb_main: invert main loop
This is an attempt to invert the main loop. Rather than spending
the majority of time executing within `gdb_main()`, turn it into
a state machine that gets polled. This will improve behaviour on
PC_HOSTED platforms as well as platforms such as ESP32 where the
main loop needs to pause occasionally.

Signed-off-by: Sean Cross <sean@xobs.io>