command: help output: use same syntax as BMDA, document all params
The syntax used in the command help was logical but unusual. Convert
it to the syntax used for the CLI help of Black Magic Debug
App (`src/platforms/hosted/cli.c`) which is similar to that of other
applications (e.g. Python `argparse`) and thus more intuitive for most
users. In particular:
- use of upper/lower-case to distinguish between keywords and
  user-supplied arguments: `keyword PARAMETER`
- `[PARAMETER]` for optional parameters (count 1)
- `PARAMETER ...` for parameters that can be repeated

There is one deviation from the syntax used by BMDA: Since the help
text for every command is just a one-liner, the default values are
documented "inline" rather than separately from the syntax.

While at it make sure all parameters are actually listed in the help
text.