target/cortexm: fix DWT mask calculation for size > 4
DWT can match power-of-two sized memory regions up to an
implementation-defined limit. Properly calculate the mask length for
region sizes of more then 4 bytes.

Note that this only fixes the calculation itself. Rather than writing
reserved bits and causing undefined behaviour we now write a properly
calculated number of mask bits to the register (up to the architecture
limit of 31). The implementation-defined limit is not queried (see
e.g. ARMv7-M Architecture Reference Manual section C1.8.1 "The DWT
comparators", heading "Checking the implemented features of the DWT
comparators") so the size of the monitored region may be less than
specified by the user.

Also no checking of the size is performed. We silently extend the
monitored region to the next power-of-two size so that the requested
range will always be included.

FWIW: In practice the "undefined behaviour" on an STM32H747 was that
no match was performed, making it seem like the memory region was not
accessed. YMMV.