imxrt: ungate boot_mode even on non-hosted
When building debug mode for hardware, the `boot_mode` flag is
consulted. However, the `boot_mode` flag is only present when
debug mode is enabled on PC_HOSTED.

Remove the requirement for this to be built on PC_HOSTED in
order to fix compiling when built on native hardware with
debug mode enabled.

Signed-off-by: Sean Cross <sean@xobs.io>