target_flash: Implemented tracking for what Flash operation is in progress and the state of each Flash block
The idea here is that flash->operation will track what state the block is in (being erased, being written, idle)
to allow special steps to be taken during preparation and conclusion of operations on the block based
on what was done to it. This allows introducing special behaviour for such cases.

Co-authored-by: dragonmux <git@dragonmux.network>
Co-authored-by: Maciej 'vesim' Kuliński <vesim809@pm.me>
Co-authored-by: Rafael Silva <perigoso@riseup.net>