samx5x: don't assume that flash block size == flash region lock size
This fixes flashing on samx5x chips with larger amounts of flash
(e.g. 512KiB).

Currently, we lock and unlock each flash region multiple times when the
flash block size differs from the size of a protected flash region --
i.e. the size of flash encoded by a single bit in the RUNLOCK register.

However, the RUNLOCK register also reads as "locked" while a write is
still in progress, and we check that before each subsequent write. When
we also manually lock after each block, and the block is smaller than a
protected region, we misinterpret our own lock as an indication that the
write is still in progress, and error out.

Co-authored-by: Katherine Temkin <k@ktemkin.com>