adiv5: use explicit AP discovery procedure for s32k3xx
These chips will hard fault if we probe for APs that are not present.  Instead,
lets detect the specific chip and explicitly initialize the APs we know it has.

Additionally, the S32K3 requires specific bits to be set in the SDA_AP registers
to allow debug access, so we explicitly initialize that as well.