platforms/stlink: add support for forcing ST-Link v2 clone detection
One of the ST-Link v2 clones I have (labelled "MINI ST-Link v2", "A"
on one side of the PCB and "QYF-0685" on the other) gets detected as
hardware version 1 rather than as a v2 clone (hardware version
257). It has PC13 pulled low via a 4.7kΩ resistor and PB11 connected
to PB10 rather than pulled up. PA15 seems to be unconnected.

It's hard to tell how other clones are wired and if PB10 is connected
to something else on them. The safest choice is to provide a
compile-time option to force "detection" of a clone (so we use PB6 for
nRST).