native/platform: fix DFU detach
DFU detach wasn't lowering the USB pullup. This meant that BMP was
rebooting into DFU bootloader mode, but the host wasn't re-enumerating
it, because it hadn't detected a detach.

In `platform_request_boot`, The `gpio_set_mode` config was for analog
instead of floating input. PA8 isn't defined as an analog pin on this
chip, so it's arguably undefined behavior.

Also confirmed that using `scb_reset_system` instead of `scb_reset_core`
to do the DFU reboot would correctly detach and enter the bootloader, if
the boot button was held down.