common/swdptap: adjust timing
Make clock cycle more symmetric, especially for `no_delay` variants.

`swdptap_seq_out_no_delay` seems to need the clock clear to be moved
back to the top of the loop, to ensure minimal delay between the falling
edge and the SWDIO change.

Also use a different loop counter that compiles to a `SUBS` `BCC`/`BCS`
pair, instead of using an explicit `CMP`.