hosted/bmda_gpiod: Add JTAG TAP init
Add routine to initialise the JTAG TAP and plumb it into the bmda jtag
init flow.

This requires that jtagtap_init is declared in jtagtap.h which
was previously conditioned on not building BMDA, this has been changed
to an unconditional declaration.

Add GPIOD probe type to the list which use jtag_scan.