hosted/gdb_if: Ensure IPv6 listening socket is used if available
It seems that on linux at least the first result from getaddr info,
when passed the AI_PASSIVE hint flag, is not always an IPv6 address.

Dual stack listening requires using an IPv6 listening socket so attempt
to find one in the result list, falling back to the first result if
none can be found.