hosted/gdb_if: Disable IPV6_V6ONLY socket option on listening socket
When IPV6_V6ONLY is not set then an IPv6 listening socket for TCP or UDP
will accept connections for both IPv4 and IPv6 allowing a single socket
to be used for both.

On linux the default value of this socket option is set by the
net.ipv6.bindv6only sysctl value, on Windows this socket option is
enabled by default and must be disabled to allow dual stack listening.

In order to ensure that the listening socket can accept connections for
both IPv6 and IPv4 if possible add code to disable this socket option
when binding to an IPv6 address, if the set fails then a warning is
given that IPv6 only listening will occur.