target: make target_reg_write and target_reg_read return size_t with 0 signalling error
The error result value for register read and write was not consistent in
the code, the individual targets would return -1 but the target wrapper
would return 0 if no single register read/write callback was available.

Additionally gdb_main did not correctly use the ssize_t type for the
result of register read/write which for read at least results in an
attempt to hexify a huge number of bytes when the read fails.

Using 0 as an error value makes this all more consistent and robust and
doesn't require use of the signed ssize_t type.