hex_uitls: Add helper function for reading integers
Reading 32 bit integers of hexdadecimal base is a common operation in
the gdb packet handling code, this often requires checking if the
integer is followed by a specific character and recording where the
end of the number was for subsequent reading.

Currently this is often accomplished with sscanf which has a significant
size footprint in the firmware and is potentially difficult to use safely.

For this reason it makes sense to have a helper function for this task.
This adds a helper function that wraps strtoul providing the following
features:
 * check if the interger has been read
 * optionally record a pointer to the end of the integer if successful
 * set the integer read by out param if successful
 * optionally check if the character immediately following the integer
   matches a given character

static inline wrappers are added for reading hex and decimal integers
respectively.