gdb_main: Convert 'v' packet handling to use exec_command table.
In preparation for removing the sscanf usage from the 'v' packet
handling convert it from one big if else if chain into a command table
with functions for each sub command.

The functions have the command prefix removed already and are passed a
const char * to the remaining packet buffer so adjustments have been made for that.