stm32l4: Fix flash for STM32U5xx
The flash size register for the STM32U5xx was in a different place,
but all the other registers were compatible with the L5.  So I
modified the way the flash size was loaded (moving the register
definition to its own location in the device structure), and then
revised the code that calls it (turns out that the only read16
call was for flash size, so I changed the name).

That was enough to get it working for U5xx, but it turns out that
the block size was also wrong.  I've corrected the block size for
U5xx, but I have not changed anything else.  However, I made a
comment for where I would suggest we start tackling that problem.

Much thanks to @ALTracer and @dragonmux for helping me get this
far so quickly.