semihosting: Esnure F-packet replies from gdb are reliably detected
During a semihosting call it is necessary to process gdb packets so
that gdb can read and write memory to perform the call as required.

This process is stopped when an F-packet is received with the result of
the call (or notification it has been abandoned before being done).

The loop in semihosting_get_gdb_response checks for an F-packet to exit
the loop however it does this by checking the first character in the
gdb packet buffer after processing the received packet.
This is incorrect as gdb_main_loop may clobber the packet buffer with
replies to gdb while processing.

Instead check if the received packet is an F-packet prior to
dispatching to gdb_main_loop for processing and remember this to exit
after it returns.