hosted: Switched out the `PC_HOSTED` macro for `CONFIG_BMDA` to indicate when we're building BMDA
The `CONFIG_` suffix has been chosen as this is kbuild-compatible and is the start of prepping the code base for a potential kbuild-base build system to replace the Makefile one.
The next several commits are designed to straighten out the configuration macros prior to adding some more for more targets.