Revert "cortexar: Read CPUID when the core is powered"
This reverts commit 9b68194a8dadeb790c03a181c85a84eecb8e73e7.

As reported on Discord, this seems to cause issues on some boards.

The ARMv7-A spec specify that: "If external debug over powerdown
is supported, these registers can be implemented in either or
both power domains."

This means that the CPUID handling on ARMv7-A is the opposite of what
ARMv8-A requires.

Signed-off-by: Mary Guillemard <mary@mary.zone>