puya: add support for PY32F002B, PY32F003, PY32F030, PY32x07x
The flashing procedure and preparatory steps are very similar across
PY32F002A, PY32F002B, PY32F003, PY32F030 and PY32x07x. There are some
differences though that need to be handled:

- The base address for the Configuration Bytes (factory calibration)
  is different between some of the families.

- On PY32x07x the Configuration Bytes are spaced at 64-bit; for all
  other families they are spaced at 32-bit.

- PY32F002B only supports 24MHz HSI and the corresponding
  Configuration Bytes are located at what would be the 4MHz
  entries (HSI_FS=0) for the other models.

- On PY32F002B some of the flash timing values are sized differently
  and EPPARA0 in particular has a different layout because of that.

- PY32F002B and PY32x07x do not have flash and RAM size stored at
  address 0x1fff0ffc and I didn't find obvious candidate so their
  sizes are hard-coded for now.

The structure and values of the DBGMCU IDCODE register are
undocumented. We use the DEV_ID / REV_ID split used by the vendor SDK
together with IDCODE values we collected ourselves. Most the
families (PY32F002A, PY32F003, PY32F030) share the same IDCODE value
of 0x60001000 but the ones that need to be handled differently do have
their own values.

For PY32F002A et al. the EPPARA4 PRETPE mask value was kept the same
as before (14 bits). The Reference Manual has conflicting information
about it: EPPARA4 26:16 (which are 11 bits) is supposed to contain
PRETPE 11:0 (which are 12 bits). FLASH_PRETPE, which is documented as
having to be initialised from the EPPARA4 value, has 14 bits for
PRETPE. The official Keil Device Family Pack v1.2.1 from Puya uses the
full upper 16 bit of EPPARA4 so presumably the reserved bits are
simply 0 and using the full 14 bit mask is probably fine.

In addition to adjusting the calibration addresses based on the model,
we now also write HSI_TRIM (using the corresponding factory
calibration value like for the flash registers). While the Reference
Manuals do not explicitly whether this is required for programming the
flash, it can make a huge difference to the accuracy of the HSI
clock (over 13% deviation were measured on one chip) and the vendor
software does the same.

Tested:
- using ST-Link v2 clone directly and via BMDA
- sequence: load, power cycle, compare, erase, power cycle, check empty
- models:
  - PY32F002AF15P6
  - PY32F002AL15S6
  - PY32F002BD15S6
  - PY32F002BF15P6
  - PY32F003L24D6
  - PY32F030F18P6
  - PY32F030F38P6
  - PY32M070K1BU7-C