gdb_main: correct read register responses when the register is not available
When reading registers, the stub may also return a string of literal ‘x’’s in place of the register data digits, to indicate that the corresponding register’s value is unavailable.
This indicates that the stub has no means to access the register contents, even though the corresponding register is known to exist.

If a register truly does not exist on the target, then it is better to not include it in the target description in the first place.