meson: enable noackmode by default
Enabling noackmode advertisement tells GDB we support and prefer noackmode, who will then enable it by default.
Right now the mode is supported but not advertised, it needs to be explicitly enabled by the user.

This should make the protocol more efficient even if by negligible amount without any drawbacks, as we have a reliable data connection making the checksums and acknowledgements redundant.