samx5x: allow specifying BOOTPROT value for `lock_bootprot`
Previously, the BOOTPROT value was hardcoded to zero which results in
reserving the entire flash space for the bootloader. This commit adds an
argument to the `lock_bootprot` command to specify the number of pages
to lock according to:

(15 - BOOTPROT) * 8 KiB      [Datasheet, Page 598]

This commit should ease flashing of Arduino-compatible binaries which
start at a 16 KiB offset (BOOTPROT = 13).